chore: move ovsx-client test code into dedicated folder
- according to coding guidelines
