+++ dev-packages/ovsx-client/src/index.ts
@@ -16,7 +16,7 @@
 
 export { OVSXApiFilter, OVSXApiFilterImpl, OVSXApiFilterProvider } from './ovsx-api-filter';
 export { OVSXHttpClient, OVSX_RATE_LIMIT } from './ovsx-http-client';
-export { OVSXMockClient } from './ovsx-mock-client';
+export { OVSXMockClient } from './test/ovsx-mock-client';
 export { OVSXRouterClient, OVSXRouterConfig, OVSXRouterFilterFactory as FilterFactory } from './ovsx-router-client';
 export * from './ovsx-router-filters';
 export * from './ovsx-types';
