+++ dev-packages/ovsx-client/src/ovsx-router-client.spec.ts
@@ -17,7 +17,7 @@
 /* eslint-disable no-null/no-null */
 
 import { OVSXRouterClient } from './ovsx-router-client';
-import { testClientProvider, registries, filterFactories } from './ovsx-router-client.spec-data';
+import { testClientProvider, registries, filterFactories } from './test/ovsx-router-client.spec-data';
 import { ExtensionLike } from './ovsx-types';
 import assert = require('assert');
